input_output_guards_funct/pyproject.toml
[project]
name = "input-output-guards-funct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.32.0",
    "openai-agents>=0.2.10",
    "python-decouple>=3.8",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "typer>=0.17.3",
    "uvicorn>=0.35.0",
]
